make IsNil guard against nil underlying errors
Almost everyone that uses quicktest for the first time uses
IsNil to check for errors, but there's a subtle mistake in doing so
because an error with a nil underlying value is not considered
to be nil for the usual `err != nil` error checking.

This change alters the semantics of `IsNil` to guard against
this case. If you really do want to check that a error-implementing
value is nil, you can still use qt.Equals.

Also add package-level docs for ContentEquals.