Fix ErrorIs for nil values
If both the LHS and the RHS are nil,
then ErrorIs should treat them as equal.

For example: c.Assert(nil, qt.ErrorIs, nil) should succeed.

If the LHS is non-nil, and the RHS is nil,
then ErrorIs should treat the RHS as an error.

For example: c.Assert(io.EOF, qt.ErrorIs, nil) should fail
due to io.EOF not being part of the error chain of nil,
rather than nil not being a valid error value.