Turn a panicking cmp.Diff into a bad check
Using Not(DeepEquals) on a type with an unexported field currently
succeeds without an error. This is because the unexported field is
raised as a panic by go-cmp, which is turned into a plain error by
the library. Not then interprets the error to mean "not equal" instead
of forwarding it to the user.

Fix this by turning any go-cmp panics into a bad check.

Fixes #153